Collect all java sources for jni_registration
* It is necessary to collect all the java source files that are reachable from jni_registration target and all the java source files (in general).
This is because the output of `jni_registration` is a cc file and a java file, From the java perspective, adding more files than required will generate
code that will never get called and removed later on. But for the cc file, adding more files than requried will generate declarations of methods that
don't have any implementation which will result in undefined symbols during linking.
* Delete cronet_aml_java and cronet_aml_api as they are no longer needed.

Bug: 304215727
Bug: 304217500
Change-Id: Iedb9d1dbbc410dc98444ee1d6c0bca392e02eeab