Add `stem` for cronet__testing
Soong does not allow two different modules to have the same output. This
did not raise an error previously because there was not a RISCV64 presubmit bot,
and `cronet__testing` was basically useless because it was getting linked with
`cronet_tests.so` until this (https://chromium-review.googlesource.com/c/chromium/src/+/4539083) landed
in 115. We have not done an import since then, Fixing this would require shipping Cronet in RISCV64 first.

So for the time being, just change the stem of the output to any other string, this should not affect the tests
as we don't have any RISCV64 tests running yet. Once we ship Cronet for RISCV64, we will revert this change.

Bug: 314275954
Change-Id: I1ca47278ef7a4ea9c394c7c5667a13b520e5ef87