Add support for CC modules depending on rust modules
All rust modules are rust_library_rlib except when a CC module tries
to depend on it. This is because Soong assumes that rust modules used
across language boundary should have their own type, in order to
align with that, create rust_ffi_static which creates a static module
that CC modules can depend on. However, rust_library_rlibs can't depend
on rust_ffi_static so we have to make two different targets.

I also added a limit on the crate name, each crate name should not
exceed 50 character, sometimes crates have significantly long names
because if a target in Chromium does not have an official "crate_name"
then the path will be used as a name. This complicates stuff a little
because AOSP assumes that every rust library crate name must be a prefix
of the module name, so when we adding two long strings, we pass the OS
limit on filenames.

Bug: 346568402
Test: mma
Change-Id: I3f561fbb4ad8ba50e68ad8434699e9ed179902e4