Relocate Android.bp destination for rust crates
We don't want to be generating Android.bp next to the BUILD.gn for rust
crates specifically as we won't be able to reach the LICENSE file that
lives next to the source code.

Generally, an assumption was made that an Android.bp (blueprint) will
live in the same directory as BUILD.gn. However, this proved to be wrong
especially with rust crates where a BUILD.gn lives in a different
directory than the srcs so we have to generate the Android.bp next to
sources. The same goes for the README.chromium where we assume it lives
next to the BUILD.gn. However, that's not the case for some targets.

This creates a disconnect where we gn2bp assumes that it will always
find a LICENSE file where it generates the Android.bp (assuming the
license generator ran correctly).

Bug: 346568402
Change-Id: Ic2e98b25b955f46b94f894ff1735b0c2191ae037

Generate correct license in Android.bp for rust crates

Bug: 346568402
Change-Id: Ib71f7951d2e829a2fdb14a98d74c82a742774aa1