Add preload API method to httpEngine
This new system API is meant to be only called from
the zygote to preload HttpEngine code into the zygote. The
current implementation preloads a single ICronetEngineBuilder class
which is used to create an underlying CronetEngine that acts as
a backend to HttpEngine.

Added a CTS test to ensure that calling the new API does not crash. It
is generally hard to create a test that ensures that the classes are
preloaded as that would require creating a custom classloader and
running the test in its own isolated environment.

Change-Id: Icb8415feb15569e199ce0cb3d8879700596e7431