Add top-level inclue_dirs anchor module
This is expected to be used once rust_bindgen modules starts being
imported from 132.0.6781.0, the problem lies in the fact that
`rust_bindgen` modules does not have `include_dirs` attribute which
means that all the of CC files include statement can't be a relative
path from the repository root which is the norm in Chromium. In order
to work around this limitation, `cronet_repository_root_include_dirs_anchor`
is introduced which does nothing but exports the repository root as
an `include_dir` to dependant modules.

Bug: 373864033
Change-Id: I265afef21a20052d64156314be5d1815bc5734f5