Add missing files to the AOSP import
These files are actually present in Chromium 126.0.6423.0, but somehow
did not make it to AOSP main during the last import. They are not even
present in upstream-import.

This commit brings in these missing files from Chromium 126.0.6423.0 to
reduce divergence. While none of these files seem to actually do
anything, it is important to eliminate this kind of divergence because
any upstream changes to these files may end up confusing Copybara and
cause spurious merge conflicts.

Test: N/A
Bug: 381421702
Change-Id: I865c27471694214f4eff30e718f371a639c86979