Improve HttpEngine creation by removing reflection
Previously, the code used reflection under the assumption that a cyclic
dependency would happened if the API depended on the Impl but that can
be fixed by having the impl depend on the stubs of the API. This means
that the Impl can only access classes that are publicly visible.

Move IHttpEngineBuilder to the Impl code as this class does not expose
any android public API (See @hide at the top of the class). If the class
has @hide, its methods are not extracted to stubs which means that it
is inaccessible via the public API, in theory this makes it an impl
class.

Bug: 380349437
Change-Id: I52a47eac797b1b78bf5a0e4b274662eaf13696cb