Add tracing to some HttpEngine's method
Added tracing to HttpEngine#Build and HttpEngine#createBuilderDelegate.

This is needed because once we start preloading the native code in
the zygote, we'll have no way to figuring out where does HttpEngine
loading code starts / end as there'll be no indication, also the tracing
sprinkled in CronetEngine does not distinguish HttpEngine from Cronet.

Change-Id: Ic396ada60a5a722d8c131ab1f9baf09c0f484c8d