Add metric to count preload attempts
This metric is using telemetry express but in this scenario, this is
fine because the caller is Zygote which is an allowed log-source. This
will be used to understand how often is HttpEngine preloaded across
different SDKs.

Test: statsd_testdrive 528 and confirmed that the metricId shows up in
the logs. I also confirmed that statsd receives the metric via some
manual changes.
Bug: 380349437

Change-Id: I9696223dcc789863711c6b3655742d124f4a5183