Import Cronet 133.0.6876.3 into AOSP
This commit brings in Cronet from Chromium 133.0.6876.3 (plus cherry-picks) into
AOSP, along with merge conflict resolutions and spot fixes.

WHAT THIS COMMIT CONTAINS

The code in this commit was imported from:

    https://chromium.googlesource.com/experimental/chromium/src/+/refs/wip/cronet/aosp-133.0.6876.3
    Commit SHA: 244475f688daad04edc5b276d2c42c83992994fc

This commit is a descendant of Chromium src tag 133.0.6876.3 (faf69aee), with
the following commits added on top:

 - 51 commits that were already backported from the Chromium src main branch
   in the previous import;
 - 9 commits newly backported from the Chromium src main branch;
 - 2 commits that are specific to 133.0.6876.3 and are not going to be submitted
   to the Chromium src main branch.

The only other changes made in this CL on top of the imported code is resolution
of merge conflicts and a spot fix for third_party/boringssl/Android.bp.

HOW TO REPRODUCE THIS COMMIT

The contents of this commit (minus the manual adjustments) can be reproduced by
running:

    gclient config --spec 'solutions = [
    {
        "name": "src",
        "url": "https://chromium.googlesource.com/experimental/chromium/src.git",
        "managed": False,
        "custom_deps": {},
        "custom_vars": {
            "checkout_copybara": True,
        },
    },
    ]
    target_os = ["android"]
    ' &&
    gclient sync --rev=refs/wip/cronet/aosp-133.0.6876.3 --shallow &&
    cd src &&
    vpython3 components/cronet/gn2bp/run_gn2bp.py

When running this command on top of this commit, the result is expected to be an
empty diff.

Fixes: 396076243
Change-Id: If9da670daa5ad64ad6a593bb245dad95a7566f0d