Remove package-private visibility divergence
Previously, we had to make some methods / classes public instead of
package-private because our testing setup was JarJaring some classes
and not jarjaring others which broke the package-visibility. We're no
longer jarjaring anything for the test classes which means that this
divergence can be removed finally.

Used aosp/3417439 and the HEAD chromium.bara.consistency as SoT for
divergence.

Fix: 394583271
Bug: 394583271
Test: TH
Change-Id: I75e8dace7b201114888a02050c27d518d46a39c4