Remove divergence from HttpEngine getters [Reland aosp/3494762]
Some APIs were introduced to HttpEngine which are not available in
Cronet. Unfortunately, they were implemented by directly introducing
those new APIs in Cronet for AOSP only but not Chromium. This introduced
divergence.

This makes the wrapper layer more complex, but keeps the complexity
localized as much as possible.

Added new CTS tests to ensure that functionlity does not break
in the future since we're now duplicating code that were previously
existing inside Cronet and was tested through the MTS.

Bug: 397215224
Test: atest NetHttpCoverageTests
Change-Id: I935fc140b919cbddc69407e802f87787f9410d58