[133.0.6876.3 v2] Indirection layer for ToT and Stable
Code outside of //external/cronet must not know about the existance of
ToT and Stable import channels, abstract them away with the use of build
system flags.

Generally speaking, we can expect //external/cronet/Android.bp to
contain most of what //external/cronet/Android.extras.bp used to contain
prior to the introducetion of ToT and Stable. It's still unclear whether
we should try to put as many things as possible into
//external/cronet/Android.bp, as oppposed to
//external/cronet/{tot, stable}/Android.extras.bp. For the time being,
we will move forward as is. Then, over time, we should converge on a
split that minimizes maintenance.

Note: since we're wiring all the targets used outside of
//external/cronet, we can get rid of the old build files.

Bug: b:396593141
Change-Id: I019011c7b9cefe2683b4b539de4ff33a6e4052b8