Add timeout option to native test runner
Previously, the test runner timeout would default to 30 minute. This was excessive especially in the cases where the test runner would hang (b/402068773) which would result in each test run hanging for 30 minutes (That's a total of 2h for our 4 tests). This will sometimes force TH to give up on the run completely and consider it stale.

The new default is 10 minutes which should be enough for our heaviest test which is net_unittests_tester takes 6minutes to pass, and it can be customized if needed.

Change-Id: Icfe6f97ddf310a4cd1b937af4adfbe41cee7ff1e