Disable SystemProxyTest#testProxyChange
This test is flaky when running in Android CI, but not in other
environments. It's unclear why and attempts to reproduce locally have
failed. Disable it for now.

I'm disabling it at the TEST_MAPPING layer because this way we only
disable it in the environment that is known to be flaky.

Bug: 432122460
Change-Id: I9e51307fb23f3e1f9c9d6fd84b5e475e4ed1896c